Remove appserver extra fields and refurbish tomcat and glassfish parsers.
Figured the as_ fields no longer are required, level information is encoded
in the syslog message and name was moved to the main table.

I've run through a set of test log files to actually see that the parsers
are working, however the Tomcat data available is very unconsistent so
the quality of the parsing is sub-optimal.